+++ packages/fury-adapter-oas3-parser/CHANGELOG.md
@@ -7,8 +7,8 @@
 - Added primitive support for 'examples' in 'Media Type Object'. The first
   example value is used for JSON media types.
 
-- Added primitive support for generating a JSON message body from a schema for
-  JSON media types. Referencing is not supported for this feature.
+- Added support for generating a JSON message body from a schema for
+  JSON media types.
 
 ### Bug Fixes